Persistence api implemented into the following classes: GameObject, Character, Player, Item, Armor, Weapon
