+++ src/main/java/com/codecool/mhmm/stickman/GameObjects/GameObject.java
@@ -1,6 +1,15 @@
 package com.codecool.mhmm.stickman.GameObjects;
 
+import javax.persistence.*;
+
+@Entity
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 public abstract class GameObject {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.SEQUENCE)
+    private long id;
+
     protected GameObjectType type;
     protected int X;
     protected int Y;
@@ -9,6 +18,10 @@ public abstract class GameObject {
         this.X = X;
         this.Y = Y;
     }
+
+    protected GameObject() {
+    }
+
     public int getX() {
         return X;
     }
@@ -24,5 +37,9 @@ public abstract class GameObject {
     public void place(int X, int Y){
         this.X = X;
         this.Y = Y;
+    }
+
+    public long getId() {
+        return id;
     }
 }
