+++ src/main/java/com/codecool/mhmm/stickman/GameObjects/Items/Item.java
@@ -1,10 +1,16 @@
 package com.codecool.mhmm.stickman.GameObjects.Items;
 
-import com.codecool.mhmm.stickman.GameObjects.Characters.Character;
+import com.codecool.mhmm.stickman.GameObjects.Characters.*;
 import com.codecool.mhmm.stickman.GameObjects.GameObject;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.persistence.Entity;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
+
+@Entity
+@Inheritance(strategy = InheritanceType.TABLE_PER_CLASS)
 public abstract class Item extends GameObject {
 
     private int value;
@@ -20,17 +26,16 @@ public abstract class Item extends GameObject {
         this.id = nextId++;
     }
 
-    void assignToCharacter(Character character) {
-        character.addItemToInventory(this);
+    protected Item() {
+    }
+
+    void assignToCharacter(Player player) {
+        player.addItemToInventory(this);
         logger.info("'{}' added to inventory", this.name);
     }
 
     public int getValue() {
         return value;
-    }
-
-    public int getId() {
-        return id;
     }
 
     public String getName() {
